+++ main.cpp
@@ -1,0 +1,8 @@
+// #include "board.h"
+#include "board.cpp"
+using namespace std;
+int main (int argc, char *argv[]) {
+  chessboard cb;
+  cb.printBoard();
+  return 0;
+}
